Refactor: delete redundant Posts model
Description:
Everytime the user finished a workout, a post would be written
to the database. These list of posts was accessed to retrieve
a single post (always the last one) to display it on profile page.
The same behaviour was achieved by creating random posts when
accessing profile_view.

Benefits: database won't get bloated with unnecessary data